fixed another dumb mistake that I shouldn't have made in the first place.

+++ Engine/Game.h
@@ -46,7 +46,7 @@ private:
 	Graphics gfx;
 	/********************************/
 	/*  User Variables              */
-	float counter = 0.0f;
+	int counter = 0;
 	int FPS = 60;
 	FrameTimer ft;
 	FrameTimer fpst;
